refactor: update count int64 (#208)
* refactor: update MockKeeper with updated type of TotalCount

related to https://github.com/edgexfoundry/go-mod-core-contracts/pull/1025

Also fixes three minor issues as scanned by golanci-lint v2


* refactor: add goVersion param into edgeXBuildGoMod for Jenkinsfile


---------

+++ internal/pkg/keeper/mock_keeper.go
@@ -104,7 +104,7 @@ func (mock *MockKeeper) Start() *httptest.Server {
 							Message:     "",
 							StatusCode:  200,
 						},
-						TotalCount: uint32(len(mock.serviceStore)),
+						TotalCount: int64(len(mock.serviceStore)),
 					},
 					Registrations: registrations,
 				}
